requirements.txt
playwright>=1.40.0
anthropic>=0.39.0
rich>=13.7.0
python-dotenv>=1.0.0
